moved the toolbar and the nav into the header
