+++ client/modules/IDE/components/Header/Toolbar.jsx
@@ -4,9 +4,9 @@ import { connect } from 'react-redux';
 import { Link } from 'react-router-dom';
 import classNames from 'classnames';
 import { withTranslation } from 'react-i18next';
-import * as IDEActions from '../actions/ide';
-import * as preferenceActions from '../actions/preferences';
-import * as projectActions from '../actions/project';
+import * as IDEActions from '../../actions/ide';
+import * as preferenceActions from '../../actions/preferences';
+import * as projectActions from '../../actions/project';
 
 import PlayIcon from '../../../images/play.svg';
 import StopIcon from '../../../images/stop.svg';
