+++ client/modules/IDE/components/Header/Toolbar.unit.test.jsx
@@ -1,7 +1,7 @@
 import React from 'react';
 import lodash from 'lodash';
 
-import { fireEvent, render, screen, waitFor } from '../../../test-utils';
+import { fireEvent, render, screen, waitFor } from '../../../../test-utils';
 import { ToolbarComponent } from './Toolbar';
 
 const renderComponent = (extraProps = {}) => {
