+++ client/modules/User/pages/DashboardView.jsx
@@ -5,7 +5,7 @@ import { withTranslation } from 'react-i18next';
 
 import browserHistory from '../../../browserHistory';
 import Button from '../../../common/Button';
-import Nav from '../../../components/Nav';
+import Nav from '../../IDE/components/Header/Nav';
 import Overlay from '../../App/components/Overlay';
 import AssetList from '../../IDE/components/AssetList';
 import AssetSize from '../../IDE/components/AssetSize';
