+++ client/modules/User/pages/SignupView.jsx
@@ -4,8 +4,8 @@ import { Helmet } from 'react-helmet';
 import { useTranslation } from 'react-i18next';
 import SignupForm from '../components/SignupForm';
 import SocialAuthButton from '../components/SocialAuthButton';
-import Nav from '../../../components/Nav';
 import RootPage from '../../../components/RootPage';
+import Nav from '../../IDE/components/Header/Nav';
 
 function SignupView() {
   const { t } = useTranslation();
